Migrate link_has_href, landmark_has_label, and disabled_attribute
