+++ lib/erblint-github/linters/github/accessibility/disabled_attribute.rb
@@ -6,23 +6,30 @@ module ERBLint
   module Linters
     module GitHub
       module Accessibility
-        class LinkHasHrefCounter < Linter
+        class DisabledAttribute < Linter
           include ERBLint::Linters::CustomHelpers
           include LinterRegistry
 
-          MESSAGE = "Links should go somewhere, you probably want to use a `<button>` instead."
+          VALID_DISABLED_TAGS = %w[button input textarea option select fieldset optgroup task-lists].freeze
+          MESSAGE = "`disabled` is only valid on #{VALID_DISABLED_TAGS.join(', ')}."
+
+          class ConfigSchema < LinterConfig
+            property :counter_enabled, accepts: [true, false], default: false, reader: :counter_enabled?
+          end
+          self.config_schema = ConfigSchema
 
           def run(processed_source)
             tags(processed_source).each do |tag|
-              next if tag.name != "a"
               next if tag.closing?
+              next if VALID_DISABLED_TAGS.include?(tag.name)
+              next if tag.attributes["disabled"].nil?
 
-              href = possible_attribute_values(tag, "href")
-              name = tag.attributes["name"]
-              generate_offense(self.class, processed_source, tag) if (!name && href.empty?) || href.include?("#")
+              generate_offense(self.class, processed_source, tag)
             end
 
-            counter_correct?(processed_source)
+            if @config.counter_enabled?
+              counter_correct?(processed_source)
+            end
           end
 
           def autocorrect(processed_source, offense)
